Pull out when one of the Sierra item merger fixtures
